feat: Add user authentication and fetch seatable data in HomeController
This commit adds user authentication logic in the HomeController's index action. If the user is signed in, their information is assigned to the `@user` instance variable. The `fetch_seatable_data` method is then called to retrieve data from the Seatable API. If the user is not signed in, a new User object is assigned to `@user`.

This change improves the functionality of the HomeController by ensuring that only signed-in users can access and fetch data from the Seatable API.